Copy ChessGuide into ChessGuideWithComments
Make a copy of the ChessGuide component, named ChessGuideWithComments.
I did this because I want to remove all reference to the comment system
from the ChessGuide component and break the comment feature into a
separate component. I think it will be easier to do that if I keep
the version with comments around for a while.