Add ChessMoveSelector to ChessGuide
I haven't found a way to test drag-and-drop behavior in test code
yet, so to get around that problem, I've added a subcomponent to
the ChessGuide component that will make it possible for tests to
select and play a user move as if the move had been drag-and-dropped.
The ChessMoveSelector component will only render if the
'renderExtraControlsForTesting' prop is set to true.