Serve each chess trap ChessGuide on its own route
Create a 'Chess Traps' page and add a link to it on the side menu. The
Chess Traps page fetches all the Chess Traps from the db and lists them
on the page as links. Each link points to a page that contains the
ChessGuide for the selected Chess Trap.