Add 'Merge Traps' page
In this commit, the 'mergeChessTrees' function is added, which merges
some number of ChessTrees together into one large tree. Some tests of
this function were also added in this commit.

The 'Merge Traps' page uses the 'mergeChessTrees' function to render
a ChessGuide component that tests a user's familiarity with multiple
chess traps at once.