Allow user to select traps they want to merge
Add user controls on the 'Merge Traps' page that allows the user to
select which traps they want to merge. There is currently a bug that
happens when the user tries to merge too many traps. This causes a
'maximum call stack size exceeded' error coming from recursive calls of
the 'mergeTwoTrees' function. In a future commit, I need to make the
'mergeTwoTrees' function non-recursive, or find some other solution.