Vary PathCompleteModal message based on state
Display different messages to the user when they complete a path. Vary
the message depending on the current GuideMode and on if all of the paths
have been completed or not. When the user completes all the paths in
learn mode, recommend switching to practice mode. When the user
completes all the paths in practice mode, suggest learning a new trap.
Otherwise, recommend resetting the board and completing the next path.