Fix: avoid unexpected modal content change
Prior to this commit, if the user completed all paths in Learn Mode
and clicked the 'Go To Practice Mode' button, the content of the modal
would change before closing. This happened because the content of the
PathCompleteModal depends on the current Mode, so when the mode changed,
the content of the modal would also change. To avoid this, I delayed the
change to Practice mode by 300 ms, to give time for the modal to close.