Fix: Don't break piece animations with arrow draws
If a prop passed to the Chessground board is changed while a piece
movement animation is playing, the animation will get interrupted. Prior
to this commit, changing the drawn arrows would sometimes interrupt these
animations, preventing some animations all together. This was most
noticable on mobile/touch devices. This commit fixes that problem.