Fix: board square highlights line up correctly
The chessground board must have a size (width === height) that is
divisible by 8 pixels, otherwise chessboard square highlighting will be
misaligned. Prior to this commit, the board was not rendering perfectly
square on narrow-width screens. I reduced the size of the board on
narrow-width screens so that the board would render with equal height
and width, and making sure that the final width and height are in fact
divisible by 8px. This fixed a problem with square highlighting
misalignment.