Add attribution prop to lessons
Add an 'attribution' typeorm entity, and include it as a column in the
chess_lessons db table. The attribution property represents a link to
the source material that was used to create the lesson. For example,
many of the lessons are based on GothamChess youtube videos. This new
field allows me to give credit to the source of the lesson. Include the
new attribution prop in server api responses to requests for lessons.