Bugfix: Hide scroll bars on Lesson pages
Prior to this commit, when navigating from any other page to a
lesson page, the vertical and horizontal scroll bars would appear
for about half a second. I think this is caused by Chessground's
setup process, but I'm not sure. Chrome devtools shows no elements
occupying the extra space, so not sure how else to address this.
My somewhat hacky solution is simply to hide the scroll bars.